Attackable has block, removed Command undo functions

+++ Assets/Resources/GameData/Scripts/FoeInfo.cs
@@ -14,6 +14,6 @@ public class FoeInfo : EntityInfo
     public FoeID foeID;
     public GameObject model;
 
-    public float health = 20f;
+    public int health = 20;
     //public List<FoeAction>
 }
